feat: Implement precise decimal arithmetic for monetary values
- Added a new utility module `money.ts` using `decimal.js` for accurate arithmetic operations on monetary values, including addition, subtraction, multiplication, division, and rounding.
- Updated `portfolio.ts` to utilize the new decimal utilities for calculating portfolio metrics, ensuring consistent precision across calculations.
- Refactored existing calculations to replace raw float operations with the new decimal methods, addressing potential floating-point inaccuracies.
- Modified backend DTOs in `CalculationDefaults`, `FreedomCalculatorResponse`, `ProjectionResult`, and `SensitivityScenario` to use `BigDecimal` instead of `double` for monetary values.
- Enhanced `FinancialFreedomService` to handle calculations with `BigDecimal`, ensuring all financial computations maintain precision.
- Added unit tests to verify the correctness of new decimal calculations and ensure existing functionality remains intact.

+++ src/main/java/org/openfinance/dto/calculator/CalculationDefaults.java
@@ -1,5 +1,6 @@
 package org.openfinance.dto.calculator;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -22,31 +23,31 @@ import lombok.NoArgsConstructor;
 public class CalculationDefaults {
 
     /** Default withdrawal rate percentage (4% rule). */
-    private double defaultWithdrawalRate;
+    private BigDecimal defaultWithdrawalRate;
 
     /** Default inflation rate percentage. */
-    private double defaultInflationRate;
+    private BigDecimal defaultInflationRate;
 
     /** Minimum allowed withdrawal rate percentage. */
-    private double minimumWithdrawalRate;
+    private BigDecimal minimumWithdrawalRate;
 
     /** Maximum allowed withdrawal rate percentage. */
-    private double maximumWithdrawalRate;
+    private BigDecimal maximumWithdrawalRate;
 
     /** Minimum allowed return rate percentage. */
-    private double minimumReturnRate;
+    private BigDecimal minimumReturnRate;
 
     /** Maximum allowed return rate percentage. */
-    private double maximumReturnRate;
+    private BigDecimal maximumReturnRate;
 
     /** Default expected annual return rate. */
-    private double defaultReturnRate;
+    private BigDecimal defaultReturnRate;
 
     /** Maximum number of projection years. */
     private int maxProjectionYears;
 
     /** Default monthly contribution amount. */
-    private double defaultMonthlyContribution;
+    private BigDecimal defaultMonthlyContribution;
 
     /** Default number of projection years. */
     private int defaultProjectionYears;
